+++ src/modules/admin/widgets/navs/NavBar.php
@@ -39,9 +39,6 @@ class NavBar extends \yii\bootstrap4\NavBar
         parent::run();
     }
 
-    /**
-     * Sorts {@see Module::$navbarItems} by key and adds home button.
-     */
     public function getItems(): array
     {
         return [...$this->getHomeItems(), ...$this->getModuleItems()];
@@ -88,9 +85,12 @@ class NavBar extends \yii\bootstrap4\NavBar
         $items = [];
 
         foreach ($i18n->getLanguages() as $language) {
+            $label = $i18n->getLabel($language);
             $items[] = [
-                'label' => '<i class="i18n-icon ' . $language . '"></i><span class="i18n-label">' . $i18n->getLabel($language) . '</span>',
-                'url' => $this->languageRoute ? Url::toRoute(array_merge($this->languageRoute, ['language' => $language])) : Url::current(['language' => $language]),
+                'label' => "<i class=\"i18n-icon$language\"></i><span class=\"i18n-label\">$label</span>",
+                'url' => $this->languageRoute
+                    ? Url::toRoute(array_merge($this->languageRoute, ['language' => $language]))
+                    : Url::current(['language' => $language]),
                 'encode' => false,
             ];
         }